refactor bagger for easier extension
