+++ archive/bagger/b_numbers.txt
@@ -1,0 +1,14 @@
+
+
+
+b2178081x - 3 smallish volumes
+
+b22454408 - short text - wl.org origin (other bucket)
+
+b2043067x - 2 images no text
+b24923333 - 2 x 550 page MM
+
+b30136155 - short MOH report - preservica bucket
+
+
+
